delete_dir check
Avoids the error that occurs when trying to delete the temp directory if
SB/CP has already deleted it.

+++ delugePostProcess.py
@@ -106,9 +106,10 @@ elif (category == categories[4]):
     log.info("Bypassing any further processing as per category.")
 
 if delete_dir:
-    try:
-        os.rmdir(delete_dir)
-        log.debug("Successfully removed tempoary directory %s." % delete_dir)
-    except:
-        log.exception("Unable to delete temporary directory.")
-    
+    if os.path.exists(delete_dir):
+        try:
+            os.rmdir(delete_dir)
+            log.debug("Successfully removed tempoary directory %s." % delete_dir)
+        except:
+            log.exception("Unable to delete temporary directory.")
+
